캐릭터들 및 아이템들 id 변경함

+++ ApplicationSoftwareProjectTeam2/network/SerializedEntity.cs
@@ -9,7 +9,7 @@ namespace ApplicationSoftwareProjectTeam2.network
     public class SerializedEntity
     {
         public byte Id { get; set; }
-        public int EntityLevel { get; set; }
+        public byte EntityLevel { get; set; }
         public int X { get; set; }
         public int Z { get; set; }
         public byte ItemId1 { get; set; }
